fix: Channel 时间戳 Long→Double（补 d4d70c2 漏提的源文件）
后端 /api/channels 的 created_at/updated_at 是 epoch 浮点数（1789614562.54），
Long 声明让 kotlinx.serialization 解码抛异常、整个频道列表静默清零。
d4d70c2 已把测试侧改成 Double，本提交补上漏掉的源文件。

+++ app/src/main/java/com/m57/hermescontrol/data/model/Channel.kt
@@ -29,8 +29,12 @@ data class Channel(
     val id: String,
     val project: String,
     val title: String,
-    val created_at: Long = 0,
-    val updated_at: Long = 0,
+    // The backend emits epoch floats (`1789614562.539754`), not integer
+    // seconds — a `Long` declaration makes kotlinx.serialization throw on
+    // decode, which silently zeroed the whole channel list (the sidebar then
+    // rendered "No project rooms yet" despite a 200 response).
+    val created_at: Double = 0.0,
+    val updated_at: Double = 0.0,
     val message_count: Int = 0,
     /**
      * Bound conversation. The desktop treats this as the room's "host" session
